--- class-components/src/components/topLine.tsx
@@ -1,15 +1,0 @@
-import { Component, ReactNode } from "react";
-import Search from "./search";
-
-
-class TopLine extends Component {
-    render(): ReactNode {
-        return (
-            <div className="topLine">
-                <Search/>
-            </div>
-        )
-    }
-}
-
-export default TopLine
